refactor trainer
