+++ ppcls/engine/eval/__init__.py
@@ -12,20 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-import os
-import sys
-__dir__ = os.path.dirname(os.path.abspath(__file__))
-sys.path.append(os.path.abspath(os.path.join(__dir__, '../')))
-
-from ppcls.utils import config
-from ppcls.engine.core import Core
-
-if __name__ == "__main__":
-    args = config.parse_args()
-    config = config.get_config(
-        args.config, overrides=args.override, show=False)
-    trainer = Core(config, mode="train")
-    trainer.train()
+from ppcls.engine.eval.classification import classification_eval
+from ppcls.engine.eval.retrieval import retrieval_eval
